Support unaligned inserts in zheap.
A new guc data_alignment has been introduced to decide the alignment
of tuple data.  The data_alignment value 0 indicates no alignment,
value 4 indicates align everything at 4 byte boundary and any other
value indicates align as per typalign of the attribute.  The main
objective of this commit is test the size of data with different
alignments.

Based on test results we might want to retain one type of alignment
and remove this GUC and associated code.