Handle cases for data_alignment = 4.
Currently for four-byte alignment, we align everything to four-byte
boundary which is not what we want for typalign 'c' or 's'.  So align the
given offset as per attalign for char and short alignment and at four-byte
boundary for other values of attalign.

Patch by me, reported by Robert Haas.