ZHeap Tuple visibility checks
Implements the basic zheap tuple visibility skeleton wherein we can
verify if the inserted tuple is visible to our command. The basic idea
used is to have few transaction slots (as of now four, needs some
testing to determine the exact number) in special space of zheap page.
Each write operation on page first needs to reserve transaction slot
in the page and update the same in tuple header and then proceed with
the actual operation. We need to reserve two bits in tuple header to
remember the information of transaction slot. There is a need of
additional bit to indicate no tranasction slot in which case tuple
will be all-visible.  This usage of additional bit will be implemented
as a separate commit.  As of now, each transaction slot contains xid,
cid and undo_rec_ptr which helps us to determine tuple visibility. We
might not need cid in transaction slot as that can be retrieved from
undo, however keeping it in page saves us from fetching undo record
pointer in many cases and due to alignment, it doesn't cost us much
space wise.