Add a 'storage_engine' reloption.
The code that decides which storage should be used for a relation.
Allowed values are 'heap' and 'zheap'. The default value for this
options is 'heap'. Also, it adds a macro 'RelationStorageIsZHeap(relation)'
which can be used to decide whether a relation uses zheap or not.

Patch by me, reviewed by Mithun C Y