Support EvalPlanQual mechanism for zheap
To support evalplanqual mechanism we need to support locking a tuple
(as of now only now only in LockTupleExclusive mode) to ensure that
nobody else can update it once the tuple is qualified by this
mechanism. We also need dirty snapshots to fetch the tuple by tupleid.
We also need to update ExecScanFetch APIs so that it can recognise
zheaptuples. We also need to adjust other visibility routines so that
they can understand locked tuples, similarly we need to adjust
zheap_update and zheap_delete to account for locked tuples.

Patch by me, reviewed and tested by Dilip.