Update the transaction id in undorecord header
Separately store the latest transaction id that has changed the tuple
in the undo record header to ensure that we don't try to process the
tuple in undo chain that is already discarded. As of now, we are
using RecentGlobalXmin to prevent it, but up-coming patches that have
Discard undo mechanism will replace it with oldest_xid_that_has_undo.