RM41618: Asynchronous Undo Management - Undo BgWorker Part.
One undoworker's launcher process which calls a loop to
periodically examine and discard undos of transactions
which are visible to all.

This is a very basic form of undoworker which can be used
by undo subsystem for discarding undos. This code will
later evolve into a more complex undoworker subsystem.