Support non-inplace-updates
Allow tuples to be updated such that a newer version of tuple will be
stored separately if any index column is updated or length of new
tuple is greater than old tuple.

For undo generation, we always generate two undo records one for the
deletion of old tuple and another for addition of new tuple.  We need
separate undo record for new tuple because during visibility checks we
sometimes need commandid and that is always stored in undo record.  To
reach new tuple from old tuple, we need ctid which is stored in old tuples
undo record.

Patch by me, review and test by Dilip Kumar and Kuntal Ghosh.