WAL log delete operation in zheap
The important consideration for logging delete operation is that we
need to generate entire tuple while performing redo operation for undo
log.  So, we are conditionally (when full page writes are off) writing
complete tuple in WAL as well.  In case full page writes are on, we
can rely on the tuple from page.

Patch by me, reviewed by Ashutosh Sharma.