Make prevlen crash safe
Currently transaction previous recrod's length is not crash safe
but there is possibility that transaction can be spread across
checkpoints, in such case while preparing the first undo record
we do need to have the length of the previous record which was
inserted before checkpoint. For fixing the same we are maintaining
this value in undo meta and wal logging this along with other
undo meta.

As part of this patch also removed one unused structure member
in undo meta.

Patch by Dilip Kumar Reviewed by Amit Kapila