Support space reuse within a page
Allow reuse of space for transactions that deletes or non-in-place
updates the tuples or updates the tuples to shorter tuples. It has
the capability to reclaim the space for deleted tuples as soon as
the transaction that has performed the operation is committed. There
is some difference between the way space is reclaimed for
transactions that are committed and all-visible vs. the transactions
that are committed but still not all-visible. In the former case, we
can just indicate in line pointer that the corresponding item is dead
whereas for later we need the capability to fetch the prior version of
tuple for transactions to which delete is not visible. As of now, we
have copied the transaction slot information in line pointer so that we
can easily reach prior version of tuple.

We try to prune the page for non-in-place updates when there is
insufficient space on the page.  The other times where we can try to
prune the page could be while Inserting if there is insufficient space
on a page or when buffer is being evicted or read, but we can do that
as a separate patch.

Mithun C Y and Amit Kapila.