Implementation of 'For Update and For Share' tuple lock modes
This requires tuple to be locked in Exclusive or Shared mode and it
will conflict update, delete and other modes of locks.  Currently,
multiple lockers for shared mode are not supported, that can be done
as a separate patch.

For other types of lock modes, user will get error "unsupported lock
mode".

Amit Kapila and Kuntal Ghosh