Make undolog meta-data checkpoints consistent.
The earlier prototype code captured undo log meta-data at an arbitrary point
in time somewhere after the redo point.  That worked only for clean shutdowns,
at which point it was consistent and correct.  To do the job properly, this
commit keeps track of two copied of each undo log's meta-data in memory: the
current meta-data, and a snapshot as of the last checkpoint.  In order to
maintain the checkpoint snapshot, every operation that modifies an undo log's
meta-data must check if we are now on the other side of a redo point.  Since
the shared memory access and lock contention would be expensive, we only
actually do that while a checkpoint is in in progress, from a moment just
before the redo point is chosen up until we discover that we are now on the
other side of a redo point, which should ideally mean that it happens only
once.

RM43038, Thomas Munro