Add a smgrsync() implementation for undofile.c.
Teach the checkpointer to go through an sgmr API call to mark a (RelFileNode,
forknum, segment) as in need of flushing to disk at the next call to
smgrsync().  Previously it called RememberFsyncRequest() in md.c directly,
but undofile.c needs to be able to participate in this scheme too.  So, add a
new function smgrrequestsync(), and have it forward to mdrequestsync() or
undofile_requestsync() as appropriate.

For now, there is a LOG message when undo segment files are fsync'd, like the
existing create/recycle/unlink messages.  These will be removed in future.

This contains a small amount of code that is copied from md.c, but the fsync
queue machinery is being redesigned so this can be rebased later to use a
common fsync queue.  See commitfest entry 18/1639 (work independent of zheap).

Thomas Munro, RM43460, reviewed by Amit Kapila