Track undo logs' is_first_rec correctly in recovery.
Previously we could get confused about whether an undo record is the first
in a transaction during recovery.  Use XLOG_UNDOLOG_ATTACH to set the flag,
since that is always emitted before the first zheap WAL record for each
transaction.

If a checkpoint happens to come between that and the first zheap operation,
it doesn't matter because then an XLOG_UNDOLOG_META will be inserted and that
will restore the is_first_rec flag, assuming it was also tracked correctly
during DO (that's a separate investigation).

Thomas Munro, RM43459, reviewed by Dilip Kumar