Make src/test/modules/test_undo compile.
After many recent changes it wasn't building.  Repair.  It isn't useful
for testing at the moment, because if you add test content to undo logs it
causes the undo worker to crash.  I need to figure out a way to prevent that
from happening and then convert this into a useful set of tests of undo log
machinery.  Watch this space.

Thomas Munro