Handle changed relfilenode while executing undo actions.
While an undo worker executes undo actions for a relation, the same
relation can be truncated. In that case, RelidByRelfilenode() when
called using the relfilenode saved in the undo record returns invalid
relation oid. Use this behaviour to figure out that the relation is
truncated, and abort the undo actions.

Because this was not handled earlier, the algorithm in
execute_undo_actions() failed to identify when exactly the undo
records switch to new pages, and kept on fetching records, without
releasing them, thus leaving behind those many records and their
buffers unreleased. This eventually leads to "no unpinned buffers
available" error in the server log. This was reproducible easily when
truncate is run immediately after interrupting a long insert, and with
shared_buffers set to minimum value.

Amit Khandekar, reviewed by Amit Kapila.
Reported by Neha Sharma.