Fix Rollback of multilockers.
Uptill now, on rollback, we never change the slot of tuple if the
multilockers flag is set on the tuple.  This is because we can't find
the next highest locker (there could be multiple lockers with same
lock level) even by traversing undo chains.  To overcome this problem,
we come up with a new design where we ensure that the tuple always
point to the transaction slot of latest inserter/updater.  For example,
say after a committed insert/update, a new request arrives to lock the
tuple in key share mode, we will keep the inserter's/updater's slot on
the tuple and set the multi-locker and key-share bit.  If the
inserter/updater is already known to be having a frozen slot (visible
to every one), we will set the key-share locker bit and the tuple will
indicate a frozen slot.  Similarly, for a new updater, if the tuple has
a single locker, then the undo will have a frozen tuple and for multi-lockers,
the undo of updater will have previous inserter/updater slot; in both cases
the new tuple will point to the updaters slot.  Now, the rollback of a single
locker will set the frozen slot on tuple and the rollback of multi-locker
won't change slot information on tuple.  We don't want to keep the slot of
locker on the tuple as after rollback, we will lose track of last
updater/inserter.

Amit Kapila, Kuntal Ghosh and Dilip Kumar