Support pruning in TPD pages.
The basic idea is process all the TPD entries in the page and remove
the old entries which are all-visible. We attempt pruning when there
is no space in the existing TPD page. Also, while accessing TPD entry,
we can consider the entry as pruned, if we find that the
ItemIdIsUnused or the block number in TPD entry is different from the
heap block number for which we are accessing the TPD entry.

Amit Kapila with help from Dilip Kumar.