Toast table support for zheap
Now, toast tables for zheap tables are created of zheap type. This imporves the
performance in terms of memory by reducing the bloat in toast tables. With zheap
type toast tables, as soon as a transaction deletes a tuple and commits, the space
can be utilised for the next insertion. Since, toast tables are larger in size
compared to ordinary tables and their updation is handled by insertion + deletion,
zheap storage is likely to benefit significantly.

Reviwed by Amit Khandekar and Amit Kapila.